requirement.txt
absl-py==2.1.0
Brotli==1.0.9
certifi==2024.2.2
charset-normalizer==2.0.4
cloudpickle==3.0.0
contourpy==1.2.0
cycler==0.12.1
filelock==3.13.1
fonttools==4.50.0
gmpy2==2.1.2
grpcio==1.62.1
gym==0.21.0
idna==3.4
Jinja2==3.1.3
kiwisolver==1.4.5
Markdown==3.6
MarkupSafe==2.1.3
matplotlib==3.8.3
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
networkx==3.1
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pillow==10.2.0
pip==24.0
protobuf==5.26.1
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
scipy==1.12.0
setuptools==57.5.0
six==1.16.0
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
torch==2.0.1
torchaudio==2.0.2
torchdiffeq==0.2.3
torchvision==0.15.2
tqdm==4.66.2
triton==2.0.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.1.0
Werkzeug==3.0.1
wheel==0.38.4
